Renamed "client-factory" -> "graphql-client-factory"
