+++ packages/create-sitecore-jss/src/templates/angular/src/app/lib/dictionary-service-factory.ts
@@ -5,7 +5,7 @@ import {
   constants,
 } from '@sitecore-jss/sitecore-jss-angular';
 import { environment as env } from '../../environments/environment';
-import { clientFactory } from './client-factory';
+import { clientFactory } from './graphql-client-factory';
 
 export class DictionaryServiceFactory {
   create(): DictionaryService {
