+++ packages/create-sitecore-jss/src/templates/angular/src/app/lib/graphql-client-factory.ts
