+++ packages/create-sitecore-jss/src/templates/angular/src/app/lib/layout-service-factory.ts
@@ -5,7 +5,7 @@ import {
   constants,
 } from '@sitecore-jss/sitecore-jss-angular';
 import { environment } from '../../environments/environment';
-import { clientFactory } from './client-factory';
+import { clientFactory } from './graphql-client-factory';
 
 export class LayoutServiceFactory {
   create(): LayoutService {
